Change NewPost & NewPostCommand's datetime to iso 8601 datetime Modify SignUpPage's title

+++ Station/Wammer.Station.Service/Station/Timeline/ResourceDownloadTask.cs
@@ -284,7 +284,8 @@ namespace Wammer.Station.Timeline
 			}
 			finally
 			{
-				File.Delete(oldFile);
+				if (File.Exists(oldFile))
+					File.Delete(oldFile);
 			}
 		}
